Changed some DAOs due to bugs

+++ src/main/java/application/service/MaterialService.java
@@ -1,5 +1,6 @@
 package application.service;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
 
 import application.dao.MaterialDao;
@@ -9,6 +10,7 @@ import application.domaim.Material;
 public class MaterialService {
 	
 	private MaterialDao dao = DaoFactory.createMaterialDao();
+	private ConcreteDesignService concreteDesignService = new ConcreteDesignService();
 	
 	public List<Material> findAll() {
 		return dao.findAll();
@@ -23,7 +25,13 @@ public class MaterialService {
 		}
 	}
 	
-	public void deleteById(Integer id) {
+	public void deleteById(Integer id) throws SQLIntegrityConstraintViolationException {
+		if (dao.findByProviderId(id).size() > 0) {
+			throw new SQLIntegrityConstraintViolationException("Hay proveedor(es) utilizando este material");
+		}
+		if (concreteDesignService.findByMaterialId(id).size() > 0) {
+			throw new SQLIntegrityConstraintViolationException("Hay dosificacion(es) utilizando este material");
+		}
 		dao.deleteById(id);
 	}
